requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.4
python-dotenv==1.1.0
anthropic==0.51.0
psycopg[binary]==3.2.7
